GHV4_1/python/requirements.txt
pyserial>=3.5
pygame-ce>=2.5.0
